analysis/requirements.txt
boto3[crt]
pyyaml == 5.*
